ui: Move Segment component to lens/common and reuse it
Change-type: patch
